Complete prototype-refactor
I went through and updated the syntax of the prototype.js file in JS3 to use the newer class syntax. It is a bit ridiculous how much easier the class syntax makes dealing with prototypal inheritance.  Glad I learned the hard way first though, so I have a better idea of what is going on.