Complete lambda-classes.js
Followed the instructions of the assignment and completed everything that was required.  Class syntax is so simple to understand and impliment.  It was almost harder to come up with filler objects for testing than it was to create the classes and their methods.  Now that this is done, I can begin working on the stretch goals for this project.